Extract addheader tests to test_main_addheader.py
Linting was failing due to a file with more than 1000 lines:

> tests/test_main.py:1:0: C0302: Too many lines in module (1032/1000)
(too-many-lines)
